backend/requirements.txt
# Core Data Science
pandas
numpy
scikit-learn
xgboost
lightgbm

# AWS & Cloud
boto3
python-dotenv

# Fairness & ML
aif360
fairlearn
mlflow

# Web Framework
fastapi
uvicorn
python-multipart
websockets

# Database
sqlalchemy
redis
psutil

# Auth & Security
PyJWT
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
email-validator==2.1.0
fastapi-users[sqlalchemy]==12.1.3

# Reporting
jinja2
reportlab

# Utilities
watchdog
deprecated

# AI Agent Dependencies (Latest Compatible Versions)
langgraph>=1.0.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-pinecone>=0.2.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
pinecone-client>=3.0.0
# sentence-transformers>=3.0.0
tiktoken>=0.7.0
openai>=1.0.0
langgraph-prebuilt==1.0.2
pinecone[grpc]>=5.0.0
# Optional (if needed)
streamlit
Python-IO

#sql
psycopg2-binary>=2.9.9